refactor: split collection file and fix cyclic dependencies
