+++ packages/datasource-mongoose/src/collections/collection.ts
@@ -1,0 +1,96 @@
+/* eslint-disable no-underscore-dangle */
+import {
+  AggregateResult,
+  Aggregation,
+  BaseCollection,
+  Caller,
+  ConditionTreeLeaf,
+  DataSource,
+  Filter,
+  PaginatedFilter,
+  Projection,
+  RecordData,
+} from '@forestadmin/datasource-toolkit';
+import { Model } from 'mongoose';
+
+import PipelineGenerator from '../utils/pipeline-generator';
+import SchemaFieldsGenerator from '../utils/schema-fields-generator';
+
+export default class MongooseCollection extends BaseCollection {
+  public readonly model: Model<RecordData>;
+
+  constructor(dataSource: DataSource, model: Model<RecordData>) {
+    super(model.modelName, dataSource);
+    this.model = model;
+    this.addFields(SchemaFieldsGenerator.buildFieldsSchema(model));
+  }
+
+  async create(caller: Caller, data: RecordData[]): Promise<RecordData[]> {
+    this.parseJSONToNestedFieldsInPlace(data);
+    const records = await this.model.insertMany(data);
+
+    const ids = records.map(record => record._id);
+    const conditionTree = new ConditionTreeLeaf('_id', 'In', ids);
+
+    return this.list(caller, new Filter({ conditionTree }), new Projection());
+  }
+
+  async list(
+    caller: Caller,
+    filter: PaginatedFilter,
+    projection: Projection,
+  ): Promise<RecordData[]> {
+    return this.model.aggregate(PipelineGenerator.find(this, this.model, filter, projection));
+  }
+
+  async update(caller: Caller, filter: Filter, patch: RecordData): Promise<void> {
+    const ids = await this.list(caller, filter, new Projection('_id'));
+    await this.model.updateMany({ _id: ids.map(record => record._id) }, patch);
+  }
+
+  async delete(caller: Caller, filter: Filter): Promise<void> {
+    const ids = await this.list(caller, filter, new Projection('_id'));
+    await this.model.deleteMany({ _id: ids.map(record => record._id) });
+  }
+
+  async aggregate(
+    caller: Caller,
+    filter: Filter,
+    aggregation: Aggregation,
+    limit?: number,
+  ): Promise<AggregateResult[]> {
+    let pipeline = PipelineGenerator.find(this, this.model, filter, aggregation.projection);
+    pipeline = PipelineGenerator.group(aggregation, pipeline);
+    if (limit) pipeline.push({ $limit: limit });
+
+    return MongooseCollection.formatRecords(await this.model.aggregate(pipeline));
+  }
+
+  protected static formatRecords(records: RecordData[]): AggregateResult[] {
+    const results: AggregateResult[] = [];
+
+    records.forEach(record => {
+      const group = Object.entries(record?._id || {}).reduce((computed, [field, value]) => {
+        computed[field] = value;
+
+        return computed;
+      }, {});
+
+      results.push({ value: record.value, group });
+    });
+
+    return results;
+  }
+
+  private parseJSONToNestedFieldsInPlace(data: RecordData[]) {
+    data.forEach(currentData => {
+      Object.entries(this.schema.fields).forEach(([fieldName, schema]) => {
+        if (schema.type === 'Column' && typeof schema.columnType === 'object') {
+          if (typeof currentData[fieldName] === 'string') {
+            currentData[fieldName] = JSON.parse(<string>currentData[fieldName]);
+          }
+        }
+      });
+    });
+  }
+}
