+++ packages/datasource-mongoose/src/index.ts
@@ -3,7 +3,7 @@ import { DataSourceFactory } from '@forestadmin/datasource-toolkit';
 
 import MongooseDatasource from './datasource';
 
-export { default as MongooseCollection } from './collection';
+export { default as MongooseCollection } from './collections/collection';
 export { default as MongooseDatasource } from './datasource';
 
 export function createMongooseDataSource(connection: Connection): DataSourceFactory {
