+++ packages/datasource-mongoose/src/utils/schema-fields-generator.ts
@@ -5,33 +5,15 @@ import {
   ColumnType,
   ManyToManySchema,
   ManyToOneSchema,
-  OneToManySchema,
   PrimitiveTypes,
   RecordData,
 } from '@forestadmin/datasource-toolkit';
 import { Model, Schema, SchemaType } from 'mongoose';
 
 import FilterOperatorBuilder from './filter-operator-builder';
-import MongooseCollection, { ManyToManyMongooseCollection } from '../collection';
 import NameFactory from './name-factory';
 
 export default class SchemaFieldsGenerator {
-  static addNewCollectionAndInverseRelationships(collections: MongooseCollection[]): void {
-    // avoid to create two times the many to many collection
-    // when iterating on the many to many type field schema
-    const createdManyToMany: string[] = [];
-    const isAlreadyCreated = name => createdManyToMany.includes(name);
-    collections.forEach(collection => {
-      Object.entries(collection.schema.fields).forEach(([fieldName, fieldSchema]) => {
-        if (fieldSchema.type === 'ManyToOne') {
-          this.addOneToManyRelation(fieldSchema, collection);
-        } else if (fieldSchema.type === 'ManyToMany' && !isAlreadyCreated(fieldName)) {
-          this.addManyToManyRelationAndCollection(fieldSchema, collection, createdManyToMany);
-        }
-      });
-    });
-  }
-
   static buildFieldsSchema(model: Model<RecordData>): CollectionSchema['fields'] {
     return Object.entries(model.schema.paths).reduce((schemaFields, [fieldName, schemaType]) => {
       const mixedFieldPattern = '$*';
@@ -107,61 +89,6 @@ export default class SchemaFieldsGenerator {
       originKeyTarget: '_id',
     } as ManyToManySchema;
     schemaFields[fieldName] = SchemaFieldsGenerator.buildColumnSchema(schema, 'String');
-  }
-
-  private static addManyToManyRelationAndCollection(
-    fieldSchema: ManyToManySchema,
-    collection: MongooseCollection,
-    createdManyToMany: string[],
-  ): void {
-    const foreignCollection = collection.dataSource.getCollection(
-      fieldSchema.foreignCollection,
-    ) as MongooseCollection;
-    const foreignKey = NameFactory.generateKey(collection.name);
-    const originKey = fieldSchema.foreignKey;
-    const { throughCollection } = fieldSchema;
-
-    const manyToMany = NameFactory.manyToManyName(
-      foreignCollection.name,
-      fieldSchema.originKey,
-      throughCollection,
-    );
-    createdManyToMany.push(manyToMany);
-
-    foreignCollection.schema.fields[manyToMany] = {
-      throughCollection,
-      foreignKey,
-      originKey,
-      foreignCollection: collection.name,
-      type: 'ManyToMany',
-      foreignKeyTarget: '_id',
-      originKeyTarget: '_id',
-    } as ManyToManySchema;
-
-    collection.dataSource.addCollection(
-      new ManyToManyMongooseCollection(
-        collection,
-        foreignCollection,
-        NameFactory.getOriginFieldNameOfIds(throughCollection),
-        manyToMany,
-      ),
-    );
-  }
-
-  private static addOneToManyRelation(
-    schema: ManyToOneSchema,
-    collection: MongooseCollection,
-  ): void {
-    const foreignCollection = collection.dataSource.getCollection(
-      schema.foreignCollection,
-    ) as MongooseCollection;
-    const field = NameFactory.oneToMany(foreignCollection.name, schema.foreignKey);
-    foreignCollection.schema.fields[field] = {
-      foreignCollection: collection.name,
-      originKey: schema.foreignKey,
-      originKeyTarget: '_id',
-      type: 'OneToMany',
-    } as OneToManySchema;
   }
 
   private static getColumnType(instance: string, field: SchemaType): ColumnType {
